# Implement Sqrt using intrinsics