--- a/base/macros.hpp
+++ b/base/macros.hpp
@@ -163,8 +163,9 @@ inline void noreturn() { std::exit(0); }
 #  error "What compiler is this?"
 #endif
 
-// We assume that the processor is at least a Nehalem.
-#define PRINCIPIA_USE_SSE2_INTRINSICS 1
+// We assume that the processor is at least a Prescott since we only support
+// 64-bit architectures.
+#define PRINCIPIA_USE_SSE3_INTRINSICS !_DEBUG
 
 // Thread-safety analysis.
 #if PRINCIPIA_COMPILER_CLANG || PRINCIPIA_COMPILER_CLANG_CL