--- a/geometry/r3_element.hpp
+++ b/geometry/r3_element.hpp
@@ -1,7 +1,7 @@
 ﻿
 #pragma once
 
-#include <nmmintrin.h>
+#include <pmmintrin.h>
 
 #include <iostream>
 #include <string>