--- a/geometry/r3_element_body.hpp
+++ b/geometry/r3_element_body.hpp
@@ -3,7 +3,7 @@
 
 #include "geometry/r3_element.hpp"
 
-#include <nmmintrin.h>
+#include <pmmintrin.h>
 
 #include <string>
 #include <type_traits>
@@ -198,7 +198,7 @@ R3Element<Scalar> operator-(R3Element<Scalar> const& right) {
 template<typename Scalar>
 R3Element<Scalar> operator+(R3Element<Scalar> const& left,
                             R3Element<Scalar> const& right) {
-#if PRINCIPIA_USE_SSE2_INTRINSICS
+#if PRINCIPIA_USE_SSE3_INTRINSICS
   return R3Element<Scalar>(_mm_add_pd(left.xy, right.xy),
                            _mm_add_sd(left.zt, right.zt));
 #else
@@ -211,7 +211,7 @@ R3Element<Scalar> operator+(R3Element<Scalar> const& left,
 template<typename Scalar>
 R3Element<Scalar> operator-(R3Element<Scalar> const& left,
                             R3Element<Scalar> const& right) {
-#if PRINCIPIA_USE_SSE2_INTRINSICS
+#if PRINCIPIA_USE_SSE3_INTRINSICS
   return R3Element<Scalar>(_mm_sub_pd(left.xy, right.xy),
                            _mm_sub_sd(left.zt, right.zt));
 #else
@@ -225,7 +225,7 @@ template<typename LScalar, typename RScalar, typename>
 R3Element<Product<LScalar, RScalar>> operator*(
     LScalar const& left,
     R3Element<RScalar> const& right) {
-#if PRINCIPIA_USE_SSE2_INTRINSICS
+#if PRINCIPIA_USE_SSE3_INTRINSICS
   __m128d const left_128d = ToM128D(left);
   return R3Element<Product<LScalar, RScalar>>(_mm_mul_pd(right.xy, left_128d),
                                               _mm_mul_sd(right.zt, left_128d));
@@ -239,7 +239,7 @@ R3Element<Product<LScalar, RScalar>> operator*(
 template<typename LScalar, typename RScalar, typename>
 R3Element<Product<LScalar, RScalar>> operator*(R3Element<LScalar> const& left,
                                                RScalar const& right) {
-#if PRINCIPIA_USE_SSE2_INTRINSICS
+#if PRINCIPIA_USE_SSE3_INTRINSICS
   __m128d const right_128d = ToM128D(right);
   return R3Element<Product<LScalar, RScalar>>(_mm_mul_pd(left.xy, right_128d),
                                               _mm_mul_sd(left.zt, right_128d));
@@ -253,7 +253,7 @@ R3Element<Product<LScalar, RScalar>> operator*(R3Element<LScalar> const& left,
 template<typename LScalar, typename RScalar, typename>
 R3Element<Quotient<LScalar, RScalar>> operator/(R3Element<LScalar> const& left,
                                                 RScalar const& right) {
-#if PRINCIPIA_USE_SSE2_INTRINSICS
+#if PRINCIPIA_USE_SSE3_INTRINSICS
   __m128d const right_128d = ToM128D(right);
   return R3Element<Quotient<LScalar, RScalar>>(_mm_div_pd(left.xy, right_128d),
                                                _mm_div_sd(left.zt, right_128d));