--- a/quantities/elementary_functions_body.hpp
+++ b/quantities/elementary_functions_body.hpp
@@ -3,6 +3,8 @@
 
 #include "quantities/elementary_functions.hpp"
 
+#include <nmmintrin.h>
+
 #include <cmath>
 #include <type_traits>
 
@@ -28,7 +30,12 @@ FORCE_INLINE(inline) Q Abs(Q const& quantity) {
 
 template<typename Q>
 SquareRoot<Q> Sqrt(Q const& x) {
+#if PRINCIPIA_USE_SSE2_INTRINSICS
+  auto const x_128d = _mm_set_sd(x / SIUnit<Q>());
+  return SIUnit<SquareRoot<Q>>() * _mm_cvtsd_f64(_mm_sqrt_sd(x_128d, x_128d));
+#else
   return SIUnit<SquareRoot<Q>>() * std::sqrt(x / SIUnit<Q>());
+#endif
 }
 
 template<typename Q>