--- a/testing_utilities/matchers.hpp
+++ b/testing_utilities/matchers.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <nmmintrin.h>
+#include <pmmintrin.h>
 
 #include <string>